update VerEq and IdEq types
